models/teamRecommenderNew/requirements.txt
flask
gensim
numpy
scikit-learn
mysql-connector-python
